requirements.txt
# Core dependencies
streamlit==1.39.0

# Language models and embeddings
torch==2.2.1
transformers==4.37.2
sentence-transformers==2.2.2
huggingface-hub==0.20.3
faiss-cpu==1.7.4

# Data handling and utilities
numpy==1.24.3
jsonschema==4.23.0
requests==2.32.3
pandas==2.2.3

# Ollama interaction
ollama==0.3.3
